Improved fetchLibrary to consider more scenarios #19112
If the library in question is already installed, as in cmake finds its
Cmake.config, the function will return and do nothing.

If the given path points to a local repository, the local repository will be
build in its current state and installed into the currents builds dist path.

If the given path is not a local path, its assumed its a remote repository, and
cmake will try to fetch the project and then build it in the given git-ref(sha,
tag or branch)

Signed-off-by: Anton Hillerband <ahille@emlix.com>